backend/pyproject.toml
[tool.poetry]
name = "mily-backend"
version = "0.1.0"
description = "Personal timeline app backend - Django REST API"
authors = ["Paul Schmidt <paulbenschmidt@gmail.com>"]
readme = "README.md"
packages = [{include = "mily"}, {include = "config"}]

[tool.poetry.dependencies]
python = "^3.12.11"
django = "^5.2.5"
djangorestframework = "^3.16.1"
django-cors-headers = "^4.7.0"
djangorestframework-simplejwt = "^5.5.1"
gunicorn = "^23.0.0"
pillow = "^11.3.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.1"
resend = "^2.16.0"
whitenoise = "^6.11.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"
poetry-plugin-export = "~1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

